分ける
