+++ src/misc/type.ts
@@ -60,46 +60,3 @@ export type ConflictPackages = ConflictPackage[];
 export interface ConflictChecker {
   checkConflict: (logicalTree: Map<string, LogicalTree>) => ConflictPackages;
 }
-
-export interface PackageUpdateInfo {
-  before: Package;
-  // もし更新がない場合はundefinedになる
-  after: Package;
-}
-
-export interface NoConflictSituation {
-  targetPackages: Package[];
-  updateTargets: PackageUpdateInfo[];
-}
-
-/**
- * Conflictを解決できないかを探す機構
- */
-export interface ConflictSolver {
-  /**
-   * conflict -> NoConflictSisutation
-   */
-  solveConflict: (
-    conflictCausePackages: Package[],
-    targetPackages: string[],
-    solveOption: { searchInRange: boolean }
-  ) => Promise<NoConflictSituation[]>;
-}
-
-interface UpdateableCheckResult {
-  updateable: boolean;
-  result: string;
-}
-
-export interface SituatuionInspectionResult {
-  target_package: string;
-  final_version: SemVer;
-  // 全体として更新して解決が可能か
-  updateable: boolean;
-  // それぞれの可否
-  update_targets_check_result: UpdateableCheckResult[];
-}
-
-export interface ConflictSolverInspector {
-  inspectSolver: (solver: NoConflictSituation) => SituatuionInspectionResult;
-}
